pyproject.toml
[project]
name = "neuropalite"
version = "0.1.0"
description = "Dyadic neurofeedback interface — Muse EEG, LSL streaming, Unity integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "muse-lsl>=2.0.2",
    "pylsl>=1.16.0",
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "mne>=1.0.0",
    "flask>=2.3.0",
    "flask-socketio>=5.3.0",
    "python-socketio>=5.9.0",
    "eventlet>=0.33.0",
    "pyyaml>=6.0",
    "pandas>=1.3.0",
    "pyxdf>=1.16.5",
    "coloredlogs>=15.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
